# online-bookstore/online-bookstore-api-rest/src/main/resources/static/openapi.yml
openapi: 3.0.4
info:
  title: Online Bookstore API
  version: 0.0.1-SNAPSHOT
  description: |-
    The Online Bookstore API provides a robust platform for managing an online book retail system.
    
    It handles book inventory management, enabling seamless tracking and updating of book stock.
    
    The API supports purchase processing and dynamic book pricing, ensuring accurate transactions and 
    flexible pricing strategies. 
    
    Additionally, it manages customer loyalty points, rewarding repeat customers with points for purchases, 
    which can be redeemed for discounts or perks, enhancing user engagement and retention.
  contact:
    email: pedraza.miguez@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /v1/orders/current:

    get:
      tags:
        - order
      summary: Get current order details
      description: Get current order details - Requires role <code>USER</code>
      operationId: getCurrentOrder
      responses:
        '200':
          description: Current order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '204':
          description: Current order is empty
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/current"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]

    delete:
      tags:
        - order
      summary: Clear current order
      description: Clear current order - Requires role <code>USER</code>
      operationId: clearCurrentOrder
      responses:
        '204':
          description: Current order cleared successfully
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/current"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]
  /v1/orders/current/books/{bookId}:

    put:
      tags:
        - order
      summary: Add book to current order
      description: Add book to current order - Requires role <code>USER</code>
      operationId: addBookToCurrentOrder
      parameters:
        - name: bookId
          in: path
          description: ID of the book to add to the current order
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 178
      requestBody:
        $ref: '#/components/requestBodies/Allocation'
      responses:
        '200':
          description: Book copies added to current order successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "BAD_REQUEST"
                message: "copies: must not be null"
                path: "/v1/orders/current/books/178"
                timestamp: "2025-04-19T02:25:11.675837"
        '409':
          description: Not enough copies in inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "CONFLICT"
                message: "Not enough copies of book with ID 178. Available: 98, Requested: 100"
                path: "/v1/orders/current/books/178"
                timestamp: "2025-04-18T16:19:34.714622"
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/current/books/178"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]

    delete:
      tags:
        - order
      summary: Remove book from current order
      description: Remove book from current order - Requires role <code>USER</code>
      operationId: removeBookFromCurrentOrder
      parameters:
        - name: bookId
          in: path
          description: ID of the book to remove from the current order
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 178
        - name: copies
          in: query
          description: Number of copies to remove from the current order
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 1000
          example: 10
      responses:
        '200':
          description: Current order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '204':
          description: Book copies removed from current order successfully
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/current/books/178"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]
  /v1/orders/current/purchase:

    post:
      tags:
        - order
      summary: Purchase current order
      description: Purchase current order - Requires role <code>USER</code>
      operationId: purchaseCurrentOrder
      requestBody:
        $ref: '#/components/requestBodies/Purchase'
      responses:
        '200':
          description: Order purchased successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              example:
                id: 178
                lines:
                  - bookId: 178
                    isbn: "9780747532743"
                    title: "Effective Java"
                    price: "21.99"
                    type: REGULAR
                    quantity: 3
                    discountRate: "0.90"
                    discountPercentage: "10.00"
                    subtotal: "59.37"
                totalPrice: "59.37"
                status: PURCHASED
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/current/purchase"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]
  /v1/orders:

    get:
      tags:
        - order
      summary: Get order history for the current customer
      description: Get order history for the current customer - Requires role <code>USER</code>
      operationId: getOrderHistory
      responses:
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]
  /v1/orders/{orderId}:

    get:
      tags:
        - order
      summary: Get order by ID
      description: Get order by ID - Requires role <code>USER</code>
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: ID of the order to retrieve
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 178
      responses:
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                status: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred"
                path: "/v1/orders/23"
                timestamp: "2025-04-18T16:19:34.714622"
      security:
        - basicAuth: [ ]
...
components:
  schemas:
    BookType:
      type: string
      description: Type of the book
      enum:
        - REGULAR
        - NEW_RELEASE
        - OLD_EDITION
...
    OrderStatus:
      type: string
      description: Status of the order
      enum:
        - CREATED
        - PURCHASED
    PaymentMethod:
      type: string
      description: Payment method used to place the order
      enum:
        - CREDIT_CARD
        - DEBIT_CARD
        - PAYPAL
        - BANK_TRANSFER
...
    PurchaseRequest:
      required:
        - paymentMethod
        - shippingAddress
      type: object
      properties:
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        shippingAddress:
          type: string
          description: Shipping address for the order
          example: "123 Main St, Springfield, IL 62704"
...
    OrderItem:
      type: object
      properties:
        bookId:
          type: integer
          format: int64
          description: ID of the book
          example: 178
        isbn:
          type: string
          description: ISBN of the book
          example: "9780747532743"
        title:
          type: string
          description: Title of the book
          example: "Effective Java"
        price:
          type: string
          format: decimal
          description: Price of the book
          example: "21.99"
        type:
          $ref: '#/components/schemas/BookType'
        quantity:
          type: integer
          format: int64
          description: Number of copies of the book in the order
          example: 3
        discountRate:
          type: string
          format: decimal
          description: Discount rate applied to the book in the order
          example: "0.80"
        discountPercentage:
          type: string
          format: decimal
          description: Discount percentage applied to the book in the order
          example: "20.00"
        subtotal:
          type: string
          format: decimal
          description: Subtotal price for the book in the order
          example: "52.77"
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 178
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalPrice:
          type: string
          format: decimal
          description: Total price of the order
          example: "219.90"
        status:
          $ref: '#/components/schemas/OrderStatus'
    Error:
      type: object
      properties:
        status:
          type: string
          description: HTTP status constant
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Path of the request
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
    Allocation:
      required:
        - copies
      type: object
      properties:
        copies:
          type: integer
          format: int64
          description: Number of copies to allocate
          minimum: 1
          maximum: 1000
          example: 2
  requestBodies:
...
    Allocation:
      description: Number of copies to allocate for a book
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Allocation'
    Purchase:
      description: Details of the purchase to be made
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PurchaseRequest'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: [ ]