# online-bookstore/online-bookstore-domain/src/main/java/es/pedrazamiguez/assessment/onlinebookstore/domain/entity/OrderItem.java
package es.pedrazamiguez.assessment.onlinebookstore.domain.entity;

import lombok.Data;

@Data
public class OrderItem {
  private Long orderId;
  private BookAllocation allocation;
  private PayableAmount payableAmount;
}